rename throttled-fetch.ts to throttledFetch.ts
