+++ packages/webapp/src/store/update.ts
@@ -13,7 +13,7 @@ import { AccountResult, Address, chains, NetworkResult } from "src/chain"
 import { subscribeToGame } from "src/store/subscriptions"
 import * as store from "src/store/atoms"
 import * as net from "src/store/network"
-import { THROTTLED, ZOMBIE } from "src/utils/throttled-fetch"
+import { THROTTLED, ZOMBIE } from "src/utils/throttledFetch"
 import { formatTimestamp, parseBigInt } from "src/utils/js-utils"
 import { GameStatus } from "src/store/types"
 import { getBlock } from "viem/actions"
